Refactor Multi-IBS integration and remove Seal Shard module
- Deleted multi-ibs-real.test.ts to streamline testing approach.
- Updated real-seal-keys.ts to use SealMultiIBSAggregator for fetching public keys.
- Refactored multi-ibs-with-seal.ts to enhance IBE signature aggregation and improve error handling.
- Removed obsolete seal-shard-with-seal.ts as it is no longer needed.
- Deleted unused test files: test_package_address.ts and test_seal_approve.ts.
