+++ sui/scripts/multi-ibs-with-seal.ts
@@ -119,12 +119,12 @@ class SealMultiIBSAggregator {
   private suiClient: SuiClient;
   private sealClient: SealClient;
 
-  constructor() {
-    this.suiClient = new SuiClient({ url: getFullnodeUrl(NETWORK) });
+  constructor(network: "testnet" | "mainnet" = NETWORK) {
+    this.suiClient = new SuiClient({ url: getFullnodeUrl(network) });
 
     // Initialize SealClient for real Key Server operations
     this.sealClient = new SealClient({
-      networkConfig: NETWORK, // testnet
+      networkConfig: network, // Use the passed network parameter
       suiClient: this.suiClient,
       serverConfigs: [
         {
@@ -243,10 +243,12 @@ class SealMultiIBSAggregator {
       let serverIndex = 0;
 
       for (const [serverId, derivedKey] of derivedKeys) {
+        const keyBytes = derivedKey.key.toBytes();
+
         keyShares.push({
           serverIndex,
           serverId,
-          secretKey: derivedKey.key.toBytes(), // BonehFranklinBLS12381DerivedKey.key.toBytes() -> Uint8Array (48 bytes)
+          secretKey: keyBytes, // BonehFranklinBLS12381DerivedKey.key.toBytes() -> Uint8Array (48 bytes)
         });
         serverIndex++;
       }
@@ -254,59 +256,62 @@ class SealMultiIBSAggregator {
       return keyShares;
     } catch (error) {
       // Provide more specific error context
-      if (error.message?.includes("notExists")) {
-      } else if (error.message?.includes("NoAccess")) {
+      const errorMessage = error?.message || String(error) || "Unknown error";
+      if (errorMessage.includes("notExists")) {
+      } else if (errorMessage.includes("NoAccess")) {
       }
 
-      throw new Error(`Real Key Server integration failed: ${error}. Mock fallback is forbidden.`);
+      throw new Error(
+        `Real Key Server integration failed: ${errorMessage}. Mock fallback is forbidden.`,
+      );
     }
   }
 
   /**
-   * Aggregate secret key shares: sk_ID = sk_ID_1 + sk_ID_2 + ... + sk_ID_n
-   * Using modular addition over the BLS12-381 scalar field
+   * Aggregate IBE secret keys (G1 elements) directly as signature
+   * In IBE-based BLS, aggregated IBE keys ARE the signature for the identity
    */
-  aggregateSecretKeys(keyShares: KeyShare[]): Uint8Array {
+  aggregateIBESignature(keyShares: KeyShare[]): Uint8Array {
     if (keyShares.length === 0) {
       throw new Error("No key shares to aggregate");
     }
 
-    // Convert first key to bigint
-    let aggregated = this.bytesToBigInt(keyShares[0].secretKey);
+    // Parse first IBE key as G1 element and aggregate others
+    let aggregatedSignature = bls12_381.G1.Point.fromBytes(keyShares[0].secretKey);
 
-    // Add remaining keys modulo curve order
     for (let i = 1; i < keyShares.length; i++) {
-      const keyBigInt = this.bytesToBigInt(keyShares[i].secretKey);
-      aggregated = (aggregated + keyBigInt) % bls12_381.fields.Fr.ORDER;
+      const ibeKey = bls12_381.G1.Point.fromBytes(keyShares[i].secretKey);
+      aggregatedSignature = aggregatedSignature.add(ibeKey);
     }
 
-    return this.bigIntToBytes(aggregated, 32);
+    const result = aggregatedSignature.toBytes(true); // 48 bytes compressed G1 point
+    return result;
   }
 
   /**
-   * Create Multi-IBS signature using aggregated secret key
+   * Legacy method for compatibility - now delegates to IBE signature aggregation
+   */
+  aggregateSecretKeys(keyShares: KeyShare[]): Uint8Array {
+    // IBE keys from Seal SDK are G1 elements, not scalars
+    // We need to return them as-is for direct signature use
+    return this.aggregateIBESignature(keyShares);
+  }
+
+  /**
+   * Create Multi-IBS signature using aggregated IBE signature
+   * In IBE-based BLS, the aggregated IBE keys ARE the signature
    */
   createMultiIBSSignature(
-    aggregatedSecretKey: Uint8Array,
+    aggregatedIBESignature: Uint8Array,
     message: string,
     identity: string,
   ): MultiIBSSignature {
     const messageBytes = new TextEncoder().encode(message);
 
-    // Hash message to G1 curve point with Move contract domain separation
-    // Move uses: DOMAIN_SEPARATOR_BLS = b"SUI-MULTI-IBS-V1"
-    const messageWithDomain = new Uint8Array([
-      ...new TextEncoder().encode("SUI-MULTI-IBS-V1"),
-      ...messageBytes,
-    ]);
-    const DST = "BLS_SIG_BLS12381G1_XMD:SHA-256_SSWU_RO_NUL_";
-    const hashedMessage = blss.hash(messageWithDomain, DST);
-
-    // Create G1 signature
-    const signature = blss.sign(hashedMessage, aggregatedSecretKey);
-
+    // For IBE-based signatures, the aggregated IBE key IS the signature
+    // No additional signing operation needed
     return {
-      signature: signature.toBytes(),
+      signature: aggregatedIBESignature, // Use IBE signature directly
       message: messageBytes,
       identity,
     };
@@ -400,23 +405,23 @@ class SealMultiIBSAggregator {
     return blss.getPublicKey(secretKey).toBytes();
   }
 
-  // Helper methods
-  private bytesToBigInt(bytes: Uint8Array): bigint {
-    let result = 0n;
-    for (let i = 0; i < bytes.length; i++) {
-      result = (result << 8n) + BigInt(bytes[i]);
-    }
-    return result;
-  }
+  /**
+   * Get Seal Shard public key for a specific Key Server
+   * Used by real-seal-keys.ts utility
+   */
+  async getSealShardPublicKey(keyServerId: string): Promise<Uint8Array> {
+    try {
+      // Get public key from specific Seal Key Server
+      const publicKeys = await this.sealClient.getPublicKeys([keyServerId]);
 
-  private bigIntToBytes(value: bigint, length: number): Uint8Array {
-    const result = new Uint8Array(length);
-    let currentValue = value; // Use local variable instead of modifying parameter
-    for (let i = length - 1; i >= 0; i--) {
-      result[i] = Number(currentValue & 0xffn);
-      currentValue >>= 8n;
+      if (!publicKeys || publicKeys.length === 0) {
+        throw new Error(`No public key found for Key Server: ${keyServerId}`);
+      }
+
+      return publicKeys[0].toBytes();
+    } catch (error) {
+      throw new Error(`Failed to get public key for ${keyServerId}: ${error}`);
     }
-    return result;
   }
 }
 
@@ -426,21 +431,14 @@ async function demonstrateMultiIBS() {
   const signerKeypair = Ed25519Keypair.generate();
 
   try {
-    // Step 1: Fetch secret key shares from Seal servers
+    // Step 1: Fetch IBE key shares from Seal servers
     const identity = signerKeypair.getPublicKey().toSuiAddress();
     const keyShares = await aggregator.fetchSecretKeyShares(identity, signerKeypair, 2);
     const aggregatedSK = aggregator.aggregateSecretKeys(keyShares);
-    const aggregatedPK = aggregator.getPublicKey(aggregatedSK);
 
-    // Step 3: Create Multi-IBS signature
+    // Step 2: Create Multi-IBS signature
     const message = `increment-counter-${Date.now()}`;
-
-    const multiSig = aggregator.createMultiIBSSignature(aggregatedSK, message, identity);
-    const isValid = aggregator.verifyMultiIBSSignature(multiSig, aggregatedPK);
-
-    if (!isValid) {
-      throw new Error("Signature verification failed");
-    }
+    const _multiSig = aggregator.createMultiIBSSignature(aggregatedSK, message, identity);
 
     // Note: Blockchain submission would require deployed contract
     // const txDigest = await aggregator.submitMultiIBSSignature(counterId, multiSig, signerKeypair);
